- Added current capacity, max capacity and height of get-all-active-saa - Created a function that gets the last time a maintenance was done (in days) and added output to get-all-active-saa endpoint fields

+++ migration/00001_init.sql
@@ -226,7 +226,7 @@ CREATE TABLE IF NOT EXISTS public.saa_type
     description character varying(1024) COLLATE pg_catalog."default" NOT NULL,
     capacity integer NOT NULL,
     diameter double precision NOT NULL,
-    height double precision NOT NULL,
+    height integer NOT NULL,
     CONSTRAINT saa_type_pkey PRIMARY KEY (id)
 );
